added utils file and function for finding the most recent logfile from a workspace.
